+++ crates/objc2/src/runtime/nsobject.rs
@@ -152,7 +152,9 @@ pub unsafe trait NSObjectProtocol {
     //
     // For example, something may have a return type of `NSString`, while
     // behind the scenes they really return `NSMutableString` and expect it to
-    // not be modified.
+    // not be modified, see [Apple's doc][apple-mut].
+    //
+    // [apple-mut]: https://developer.apple.com/library/archive/documentation/General/Conceptual/CocoaEncyclopedia/ObjectMutability/ObjectMutability.html
 
     /// Check if the object is an instance of a specific class, without
     /// checking subclasses.